A list object for aquifer parameters

Create an object of class 'aquifer'

aquifer_param(
  input_unit = NULL,
  output_unit = NULL,
  model = NULL,
  flow_type = NULL,
  water_drive = NULL,
  phi = NULL,
  perm_h = NULL,
  perm_v = NULL,
  h_a = NULL,
  r_a = NULL,
  r_R = NULL,
  w_a = NULL,
  l_a = NULL,
  tetha = NULL,
  mu_water = NULL,
  c_water = NULL,
  c_rock = NULL,
  pressure = NULL
)

Arguments

input_unit
a unit system for parameters, a character string either 'SI' or 'Field'

output_unit
a unit system for properties, a character string either 'SI' or 'Field'

model
state of flow in the aquifer, a character string either 'uss' for the un-steady state flow or 'pss' for the pseudo-steady state flow

flow_type
a character string either 'radial' or 'linear'

water_drive
a character string either 'edge' or 'bottom'

phi
aquifer porosity, a numeric fraction

perm_h
aquifer horizontal permeability in 'md' in both 'SI' and 'Field' input unit systems. A NULL value must be used for the combination of 'uss', 'linear', and 'bottom' flow

perm_v
aquifer vertical permeability in 'md' in both 'SI' and 'Field' input unit systems. A NULL value must be used for the combination of 'uss', 'linear', 'edge' flow. A NULL value must be used for the combination of 'uss', 'radial', 'edge' flow. A NULL value must be used for the combination of 'pss', 'radial', 'edge' flow.

h_a
aquifer height in 'm' or 'ft' in 'SI' and 'Field' input unit systems, respectively.

r_a
aquifer radius in 'm' or 'ft' in 'SI' and 'Field' input unit systems, respectively. A NULL value must be used for the combination of 'uss', 'linear', 'edge' flow. A NULL value must be used for the combination of 'uss', 'linear', 'bottom' flow.

r_R
reservoir radius in 'm' or 'ft' in 'SI' and 'Field' input unit systems, respectively. A NULL value must be used for the combination of 'uss', 'linear', 'edge' flow. A NULL value must be used for the combination of 'uss', 'linear', 'bottom' flow.

w_a
aquifer width in 'm' or 'ft' in 'SI' and 'Field' input unit systems, respectively. A NULL value must be used for the combination of 'uss', 'radial', 'edge' flow. A NULL value must be used for the combination of 'uss', 'radial', 'bottom' flow. A NULL value must be used for the combination of 'pss', 'radial', 'edge' flow.

l_a
aquifer length in 'm' or 'ft' in 'SI' and 'Field' input unit systems, respectively. A NULL value must be used for the combination of 'uss', 'radial', 'edge' flow. A NULL value must be used for the combination of 'uss', 'radial', 'bottom' flow. A NULL value must be used for the combination of 'pss', 'radial', 'edge' flow.

tetha
fraction of reservoir encircled by the aquifer, reported in "degrees" in both 'SI' and 'Field' input unit systems. A NULL value must be used for the combination of 'uss', 'radial', 'bottom' flow. A NULL value must be used for the combination of 'uss', 'linear', 'edge' flow. A NULL value must be used for the combination of 'uss', 'linear', 'bottom' flow.

mu_water
water viscosity in 'mPa.s' or 'cp' in 'SI' and 'Field' input unit systems, respectively

c_water
water compressibility in '1/kPa' or '1/psi' in 'SI' and 'Field' input unit systems, respectively

c_rock
rock compressibility in '1/kPa' or '1/psi' in 'SI' and 'Field' input unit systems, respectively

pressure
a numeric vector of pressure data at the boundary of reservoir/aquifer. Must have the same length as the 'aquifer_time()' object

Value

a list of class 'aquifer' with all the required parameters for the aquifer_predict() S3 methods

Examples

aquifer_param_01 <- aquifer_param(input_unit = "Field", output_unit = "Field",
model = "uss", flow_type = "radial", water_drive = "edge", phi = 0.2, perm_h = 100,
h_a = 47, r_a = 2e4, r_R = 2e3, tetha = 360, mu_water = 0.34, c_water = 4e-6,
c_rock = 3e-6, pressure = c(3456, 3425, 3387, 3350, 3312))

aquifer_param_01

#> $input_unit
#> [1] "Field"
#> 
#> $output_unit
#> [1] "Field"
#> 
#> $model
#> [1] "veh_rad_edge"
#> 
#> $phi
#> [1] 0.2
#> 
#> $perm_h
#> [1] 100
#> 
#> $h_a
#> [1] 47
#> 
#> $r_a
#> [1] 20000
#> 
#> $r_R
#> [1] 2000
#> 
#> $tetha
#> [1] 360
#> 
#> $mu_water
#> [1] 0.34
#> 
#> $c_water
#> [1] 4e-06
#> 
#> $c_rock
#> [1] 3e-06
#> 
#> $pressure
#> [1] 3456 3425 3387 3350 3312
#> 
#> attr(,"class")
#> [1] "veh_rad_edge" "aquifer"     

aquifer_param_02 <- aquifer_param(input_unit = "SI", output_unit = "SI",
model = "uss", flow_type = "radial", water_drive = "bottom", phi = 0.2, perm_h = 100,
perm_v = 25, h_a = 25, r_a = 6000, r_R = 600, mu_water = 0.34, c_water = 6e-7,
c_rock = 4.5e-7, pressure = c(3456, 3425, 3387, 3350, 3312) * 6.895)

aquifer_param_02

#> $input_unit
#> [1] "SI"
#> 
#> $output_unit
#> [1] "SI"
#> 
#> $model
#> [1] "ykh_rad_bottom"
#> 
#> $phi
#> [1] 0.2
#> 
#> $perm_h
#> [1] 100
#> 
#> $perm_v
#> [1] 25
#> 
#> $h_a
#> [1] 25
#> 
#> $r_a
#> [1] 6000
#> 
#> $r_R
#> [1] 600
#> 
#> $mu_water
#> [1] 0.34
#> 
#> $c_water
#> [1] 6e-07
#> 
#> $c_rock
#> [1] 4.5e-07
#> 
#> $pressure
#> [1] 23829.12 23615.38 23353.36 23098.25 22836.24
#> 
#> attr(,"class")
#> [1] "ykh_rad_bottom" "aquifer"       

aquifer_param_03 <- aquifer_param(input_unit = "Field", output_unit = "Field",
model = "pss", flow_type = "radial", water_drive = "edge", phi = 0.2, perm_h = 100,
h_a = 47, r_a = 2e4, r_R = 2e3, tetha = 360, mu_water = 0.34, c_water = 4e-6,
c_rock = 3e-6, pressure = c(3456, 3425, 3387, 3350, 3312))

aquifer_param_03

#> $input_unit
#> [1] "Field"
#> 
#> $output_unit
#> [1] "Field"
#> 
#> $model
#> [1] "fetk_rad_edge"
#> 
#> $phi
#> [1] 0.2
#> 
#> $perm_h
#> [1] 100
#> 
#> $h_a
#> [1] 47
#> 
#> $r_a
#> [1] 20000
#> 
#> $r_R
#> [1] 2000
#> 
#> $tetha
#> [1] 360
#> 
#> $mu_water
#> [1] 0.34
#> 
#> $c_water
#> [1] 4e-06
#> 
#> $c_rock
#> [1] 3e-06
#> 
#> $pressure
#> [1] 3456 3425 3387 3350 3312
#> 
#> attr(,"class")
#> [1] "fetk_rad_edge" "aquifer"      

aquifer_param_04 <- aquifer_param(input_unit = "Field", output_unit = "Field",
model = "uss", flow_type = "linear", water_drive = "edge", phi = 0.2, perm_h = 100,
h_a = 47, w_a = 30000, l_a = 10000, mu_water = 0.34, c_water = 4e-6,
c_rock = 3e-6, pressure = c(3456, 3425, 3387, 3350, 3312))

aquifer_param_04

#> $input_unit
#> [1] "Field"
#> 
#> $output_unit
#> [1] "Field"
#> 
#> $model
#> [1] "nb_lin_edge"
#> 
#> $phi
#> [1] 0.2
#> 
#> $perm_h
#> [1] 100
#> 
#> $h_a
#> [1] 47
#> 
#> $w_a
#> [1] 30000
#> 
#> $l_a
#> [1] 10000
#> 
#> $mu_water
#> [1] 0.34
#> 
#> $c_water
#> [1] 4e-06
#> 
#> $c_rock
#> [1] 3e-06
#> 
#> $pressure
#> [1] 3456 3425 3387 3350 3312
#> 
#> attr(,"class")
#> [1] "nb_lin_edge" "aquifer"    

aquifer_param_05 <- aquifer_param(input_unit = "Field", output_unit = "Field",
model = "uss", flow_type = "linear", water_drive = "bottom", phi = 0.2, perm_v = 10,
h_a = 47, w_a = 4000, l_a = 4000, mu_water = 0.34, c_water = 4e-6,
c_rock = 3e-6, pressure = c(3456, 3425, 3387, 3350, 3312))

aquifer_param_05

#> $input_unit
#> [1] "Field"
#> 
#> $output_unit
#> [1] "Field"
#> 
#> $model
#> [1] "nb_lin_bottom"
#> 
#> $phi
#> [1] 0.2
#> 
#> $perm_v
#> [1] 10
#> 
#> $h_a
#> [1] 47
#> 
#> $w_a
#> [1] 4000
#> 
#> $l_a
#> [1] 4000
#> 
#> $mu_water
#> [1] 0.34
#> 
#> $c_water
#> [1] 4e-06
#> 
#> $c_rock
#> [1] 3e-06
#> 
#> $pressure
#> [1] 3456 3425 3387 3350 3312
#> 
#> attr(,"class")
#> [1] "nb_lin_bottom" "aquifer"