Generic function for cumulative water influx predictions

Generate a data frame of cumulative water influx estimates according to the class of 'aquifer_lst' and 'time_lst' objects

aquifer_predict(aquifer_lst, time_lst)

Arguments

aquifer_lst
a list object of class 'aquifer'

time_lst
a list object of class 'time'

Value

a data frame of cumulative water influx estimates according to the class of 'aquifer_lst' and 'time_lst' objects

References

Yildiz T, Khosravi A (2007). “An Analytical Bottomwaterdrive Aquifer Model for Material-Balance Analysis.” SPE Reservoir Evaluation \& Engineering, 10(06), 618--628. ISSN 1094-6470, doi: 10.2118/103283-PA , https://doi.org/10.2118/103283-PA.

Nabor GW, Barham RH (1964). “Linear Aquifer Behavior.” Journal of Petroleum Technology, 16(05), 561--563. ISSN 0149-2136, doi: 10.2118/791-PA , https://doi.org/10.2118/791-PA.

Fetkovich MJ (1971). “A Simplified Approach to Water Influx Calculations-Finite Aquifer Systems.” Journal of Petroleum Technology, 23(07), 814--828. ISSN 0149-2136, doi: 10.2118/2603-PA , https://doi.org/10.2118/2603-PA.

Van Everdingen AF, Hurst W (1949). “The Application of the Laplace Transformation to Flow Problems in Reservoirs.” Journal of Petroleum Technology, 1(12), 305--324. ISSN 0149-2136, doi: 10.2118/949305-G , https://doi.org/10.2118/949305-G.

Examples

aquifer_time_1 <- aquifer_time(c(0:4) * 365, unit = "day")
aquifer_param_01 <- aquifer_param(input_unit = "Field", output_unit = "Field",
model = "uss", flow_type = "radial", water_drive = "edge", phi = 0.2, perm_h = 100,
h_a = 47, r_a = 2e4, r_R = 2e3, tetha = 360, mu_water = 0.34, c_water = 4e-6,
c_rock = 3e-6, pressure = c(3456, 3425, 3387, 3350, 3312))
results_01 <- aquifer_predict(aquifer_param_01, aquifer_time_1)

results_01

#>         Date Time (days) We (MMbbl)
#> 1 2020-07-04           0  0.0000000
#> 2 2021-07-04         365  0.1764468
#> 3 2022-07-04         730  0.6069313
#> 4 2023-07-04        1095  1.1268676
#> 5 2024-07-03        1460  1.6674785